Fix button type detection for IE lte 8
`pathname` in IE lte 9 does not begin with '/', thus a '/' is prepended
before `pathname` is splitted by `/\/+/`. However, IE lte 8 have a bug
where splitting by regex would ignore empty strings in result. The fix
here is to prepend ' /' (with an extra space), so that the first field
will always be non empty.